asker/requirement.txt
requests
mysql_connector_python